增加硬件输入控制器插件

+++ src/Interface/IEventService.h
@@ -49,9 +49,9 @@ public:
 	virtual void PostEvent( XE::real dt, EventPtr val ) = 0;
 
 public:
-	virtual XE::uint64 RegisterListener( XE::uint32 event, ListenerType listener ) = 0;
+	virtual XE::uint64 RegisterListener( XE::uint64 event, ListenerType listener ) = 0;
 
-	virtual void UnregisterListener( XE::uint32 event, XE::uint64 index ) = 0;
+	virtual void UnregisterListener( XE::uint64 event, XE::uint64 index ) = 0;
 
 };
